moved new i2c driver to a seperate file
